+++ CHANGELOG.md
@@ -23,6 +23,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![API] Rename field `SdmxWebSource#description` as `SdmxWebSource#name`
 - ![API] Invert fields in `DataDetail`
 - ![API] Set field `Dataflow#description` as optional
+- ![API] Set default values to `DataQuery`
 - ![FORMAT] Improve parsing of time formats
 - ![FORMAT] Rename module `sdmx-dl-format-util ` as `smdx-dl-format-base`
 - ![FORMAT] Rename module `sdmx-dl-provider-util ` as `smdx-dl-provider-base`